requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
attrs==23.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
cssselect2==0.7.0
Django==4.1.3
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-spectacular==0.26.2
future==0.18.2
html5lib==1.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
PyPDF3==1.0.6
pyrsistent==0.19.3
python-bidi==0.4.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
twilio==7.15.4
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.8
